pyproject.toml
[project]
name = "macos-terminal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
]
